reformat codes and add UT test.
